+++ src/podofo/doc/PdfFontCID.h
@@ -81,7 +81,7 @@ class PdfFontCID : public PdfFont {
     void EmbedFont() override;
 
     void EmbedSubsetFont() override;
-    void AddUsedSubsettingGlyphs (const PdfString &sText, long lStringLen) override;
+    void AddUsedSubsettingGlyphs (const PdfString &sText, size_t lStringLen) override;
 
  private:
     /** Create the DW and W entries which contain
